backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tailored-job-application"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "google-genai>=1.0.0",
    "python-multipart>=0.0.12",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
dev = ["ruff", "pytest", "pytest-asyncio", "httpx"]

[tool.hatch.build.targets.wheel]
packages = ["app"]
